Add files via upload
his is the Home page which i created for our Group Project. The home page contain a background image which i added using src and a front image. This is where the about, which is our Nam Entertainment Group Project page description is. There is also a Show page that displays recap of current and previus event through photos. Using JavaScript, i created the slideshows in the Shows.html.