Add files via upload
home is the Home page of Nam entertainment  i created. It includes the home page photos and navigations links to the other four pages. It's also were details of Nam Entertainment is explained in the about section. Additionally, there's a Shows page that details the current and past events that occurs. it's all being displayed in forms of pictures. The Shows page contains Fashion Shows, Music Concerts, Sports, Comedy Shows and Promotions of up coming events. The Shows page have images which are in slideshows and Shows detailed in ordered list.